# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vantor API",
    "version": "0.8.0-one-stop"
  },
  "paths": {
...
    "/api/v1/requisitions": {
      "post": {
        "tags": [
          "purchase"
        ],
        "summary": "Create Req",
        "operationId": "create_req_api_v1_requisitions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReqIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Req Api V1 Requisitions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/requisitions/{rid}/submit": {
      "post": {
        "tags": [
          "purchase"
        ],
        "summary": "Submit Req",
        "operationId": "submit_req_api_v1_requisitions__rid__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Submit Req Api V1 Requisitions  Rid  Submit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders": {
      "post": {
        "tags": [
          "purchase"
        ],
        "summary": "Create Po",
        "operationId": "create_po_api_v1_purchase_orders_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Po Api V1 Purchase Orders Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "purchase"
        ],
        "summary": "List Pos",
        "operationId": "list_pos_api_v1_purchase_orders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Cursor"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Pos Api V1 Purchase Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders/{pid}/approve": {
      "post": {
        "tags": [
          "purchase"
        ],
        "summary": "Approve Po",
        "operationId": "approve_po_api_v1_purchase_orders__pid__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Approve Po Api V1 Purchase Orders  Pid  Approve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders/{pid}/send": {
      "post": {
        "tags": [
          "purchase"
        ],
        "summary": "Send Po",
        "operationId": "send_po_api_v1_purchase_orders__pid__send_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Send Po Api V1 Purchase Orders  Pid  Send Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders/{pid}/receipts": {
      "post": {
        "tags": [
          "purchase"
        ],
        "summary": "Receive",
        "operationId": "receive_api_v1_purchase_orders__pid__receipts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Receive Api V1 Purchase Orders  Pid  Receipts Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders/{pid}/invoices": {
      "post": {
        "tags": [
          "purchase"
        ],
        "summary": "Create Invoice",
        "operationId": "create_invoice_api_v1_purchase_orders__pid__invoices_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Invoice Api V1 Purchase Orders  Pid  Invoices Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoices/{iid}/approve": {
      "post": {
        "tags": [
          "purchase"
        ],
        "summary": "Approve Invoice",
        "operationId": "approve_invoice_api_v1_invoices__iid__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "iid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Approve Invoice Api V1 Invoices  Iid  Approve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders/{pid}": {
      "get": {
        "tags": [
          "purchase"
        ],
        "summary": "Get Po",
        "operationId": "get_po_api_v1_purchase_orders__pid__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Po Api V1 Purchase Orders  Pid  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PoIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2,
            "title": "Code"
          },
          "supplier_id": {
            "type": "string",
            "minLength": 1,
            "title": "Supplier Id"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": ""
          },
          "category_id": {
            "type": "string",
            "title": "Category Id",
            "default": ""
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PoLineIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "code",
          "supplier_id",
          "lines"
        ],
        "title": "PoIn"
      },
...
      "ReceiptIn": {
        "properties": {
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ReceiptLineIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "lines"
        ],
        "title": "ReceiptIn"
      },
...
      "ReqIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2,
            "title": "Code"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 2,
            "title": "Title"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ReqLineIn"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title"
        ],
        "title": "ReqIn"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}